Added extract_first_layer
This function will extract the "first layer" of the Music Tree:
This includes the staff size, the page size, the header
information, etc.

Added statement.hpp in order to facilitate the usage of
statements in nodes: for now, there is only the AbstractStatement
and the Statement class. Added statement.hpp to the CMakeLists
file.